refactor: delete 7K lines dead code — Marblestone, CUDA cone/Poincaré, entity model, and gate disabled embedders
Remove dead modules tied to the deprecated Memex/Marblestone layer:
- Delete marblestone/ module (~740 lines): inline EdgeType, Domain, NeurotransmitterWeights into graph_edge
- Delete CUDA cone/ and poincare/ modules (~3,700 lines): unused hyperbolic geometry kernels + CPU fallbacks
- Delete pretrained entity/ model (~1,687 lines): dead KEPLER E11 implementation (E11 weight=0)
- Delete MCP middleware/ module (unused)
- Gate E2/E3/E4 inference behind TEMPORAL_EMBEDDERS_ENABLED (skip GPU when weight=0)
- Gate E11 inference behind E11_ENTITY_ENABLED (skip GPU when weight=0)
- Clean unused dependencies from core, embeddings, storage Cargo.toml

72 files changed, 459 insertions, 7,035 deletions.

+++ crates/context-graph-benchmark/src/runners/causal.rs
@@ -688,14 +688,14 @@ impl CausalBenchmarkRunner {
         for pair in dataset.pairs.iter().take(total) {
             // Compute forward (cause→effect) and reverse (effect→cause) similarities
             let forward = compute_asymmetric_similarity(
-                pair.strength as f32,
+                pair.strength,
                 CausalDirection::Cause,
                 CausalDirection::Effect,
                 None,
                 None,
             );
             let reverse = compute_asymmetric_similarity(
-                pair.strength as f32,
+                pair.strength,
                 CausalDirection::Effect,
                 CausalDirection::Cause,
                 None,
@@ -787,7 +787,7 @@ impl CausalBenchmarkRunner {
         for pair in &dataset.pairs {
             // Symmetric means: compute_asymmetric_similarity with Unknown direction
             let sim = compute_asymmetric_similarity(
-                pair.strength as f32,
+                pair.strength,
                 CausalDirection::Unknown,
                 CausalDirection::Unknown,
                 None,
@@ -810,7 +810,7 @@ impl CausalBenchmarkRunner {
         for pair in &dataset.pairs {
             // cause→effect direction modifiers enabled
             let sim = compute_asymmetric_similarity(
-                pair.strength as f32,
+                pair.strength,
                 CausalDirection::Cause,
                 CausalDirection::Effect,
                 None, // No intervention context
@@ -837,7 +837,7 @@ impl CausalBenchmarkRunner {
             };
 
             let sim = compute_asymmetric_similarity(
-                pair.strength as f32,
+                pair.strength,
                 CausalDirection::Unknown, // No direction modifier
                 CausalDirection::Unknown,
                 Some(&intervention),
